requirements.txt
langchain
openai
streamlit
fastapi
uvicorn
faiss-cpu
python-dotenv
mlflow
langchain-community
tiktoken
pypdf
langchain-openai